[feat]: 버튼 이벤트 구현 class 세분화 keyAction 인터페이스 생성 keyAction 인터페이스를 상속받아서 action 클래스 구현 진입점 mod 재설정
