+++ src/main/kotlin/org/minecrazy/Config.kt
@@ -8,7 +8,7 @@ import net.neoforged.fml.common.EventBusSubscriber
 import net.neoforged.fml.event.config.ModConfigEvent
 import net.neoforged.neoforge.common.ModConfigSpec
 // 예제에 있는 Config임
-@EventBusSubscriber(modid = ExampleMod.MOD_ID, bus = EventBusSubscriber.Bus.MOD)
+@EventBusSubscriber(modid = MinecrazyMod.MOD_ID, bus = EventBusSubscriber.Bus.MOD)
 object Config {
 
     private val builder = ModConfigSpec.Builder()
