+++ src/main/kotlin/org/minecrazy/KeyAction.kt
@@ -1,0 +1,5 @@
+package org.minecrazy
+
+interface KeyAction {
+    fun execute()
+}
